requirements.txt
disnake
python-decouple
